rpm/macros.d: Make disabling Go modules configurable for %gobuild and %gotest
Having this unconditionally disabled without some way to change it
causes more heartache for downstream users than necessary.

It doesn't hurt to make it default to disable it while providing a
way for downstream users who need the Go module features (e.g. in COPR)
to be able to use them.