Add RPM generator for bundled Provides
Any modules.txt file within a subdirectory of /usr/share/licenses will
be processed. While this isn't strictly a license file, it contains
information about the other sources that are included in the package.
Packages can include this file alongside the license files from the
vendored libraries as a summary of sorts. Alternatively, the modules.txt
can be marked with `%ghost`, which is sufficient for the generator.
There are not any other good places to install these files. %_docdir is
not an option, as generators do not search in that directory.